fix(elevator): demo拆解与规范
